+++ src/textual_filelink/__init__.py
@@ -3,6 +3,15 @@ from .file_link import FileLink
 from .file_link_list import FileLinkList
 from .file_link_with_icons import FileLinkWithIcons
 from .icon import Icon
-from .utils import sanitize_id
+from .utils import format_duration, format_time_ago, sanitize_id
 
-__all__ = ["FileLink", "FileLinkWithIcons", "CommandLink", "FileLinkList", "Icon", "sanitize_id"]
+__all__ = [
+    "FileLink",
+    "FileLinkWithIcons",
+    "CommandLink",
+    "FileLinkList",
+    "Icon",
+    "sanitize_id",
+    "format_duration",
+    "format_time_ago",
+]